Refactor of enki
Reworked sp_dev sp_id sp_net sp_topic with clear separation
between the classes.
Added callbacks and a UI protocol class to separate the
display to the user from the core of the app.
Moved the shell interface to a class that follows that protocol
and is not in the main file anymore.
Removed the logger commands.

Signed-off-by: Pierre Fayolle <pierre.fayolle@vossloh.com>